+++ sdk/storage/azure_storage_blob/tests/block_blob_client.rs
@@ -431,6 +431,7 @@ async fn upload_empty(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
+// This test generates a large recording, so marking as live-only.
 #[recorded::test(live)]
 async fn upload_large(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     let stage_block_count = Arc::new(AtomicUsize::new(0));
@@ -474,11 +475,12 @@ async fn upload_large(ctx: TestContext) -> Result<(), Box<dyn Error>> {
         expected_stage_block_count
     );
 
+    container_client.delete(None).await?;
+
     Ok(())
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_commit_block_list_content_headers(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -533,7 +535,6 @@ async fn test_commit_block_list_content_headers(ctx: TestContext) -> Result<(), 
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_get_block_list_types(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -592,7 +593,6 @@ async fn test_get_block_list_types(ctx: TestContext) -> Result<(), Box<dyn Error
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_stage_block_transactional_checksums(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -676,7 +676,6 @@ async fn test_stage_block_transactional_checksums(ctx: TestContext) -> Result<()
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_commit_block_list_with_tags(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -709,7 +708,6 @@ async fn test_commit_block_list_with_tags(ctx: TestContext) -> Result<(), Box<dy
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_stage_block_from_url_source_if_match(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -771,7 +769,6 @@ async fn test_stage_block_from_url_source_if_match(ctx: TestContext) -> Result<(
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_upload_blob_from_url_source_if_match(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -827,7 +824,6 @@ async fn test_upload_blob_from_url_source_if_match(ctx: TestContext) -> Result<(
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_upload_block_blob_with_tags(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();
@@ -855,7 +851,6 @@ async fn test_upload_block_blob_with_tags(ctx: TestContext) -> Result<(), Box<dy
 }
 
 #[recorded::test]
-#[ignore = "need to investigate live test pipeline failures"]
 async fn test_upload_blob_from_url_source_timestamp_conditions(
     ctx: TestContext,
 ) -> Result<(), Box<dyn Error>> {